# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TODO API Documentation",
        "description": "TODO API documentation",
        "contact": {
            "email": "aleksey.derevyanko81@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/todos": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "Get auth user todos",
                "description": "Index",
                "operationId": "5510844e30a45dbdcd411af1d4e00991",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Amount items per page",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "3"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page index",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "payload": {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "todo": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number",
                                                                    "example": 2
                                                                },
                                                                "title": {
                                                                    "type": "string",
                                                                    "example": "Title"
                                                                },
                                                                "description": {
                                                                    "type": "string",
                                                                    "example": "Some description"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "example": "2024-01-30 06:06:17"
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "example": "2024-01-30 06:06:17"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Todo"
                ],
                "summary": "Store todo",
                "description": "Store",
                "operationId": "89cb70495ecb1d8391816b58da827409",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "New todo title",
                                    "description": "New todo description"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": 201
                                        },
                                        "payload": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 2
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Title"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Some description"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-01-30 06:06:17"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-01-30 06:06:17"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The title field is required"
                                            },
                                            "example": [
                                                {
                                                    "message": "The title field is required"
                                                },
                                                {
                                                    "message": "The title field must not be more than 255 characters"
                                                },
                                                {
                                                    "message": "The title field must not be less than 3 characters"
                                                }
                                            ]
                                        },
                                        "description": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The description field must not be more than 1000 characters"
                                            },
                                            "example": [
                                                {
                                                    "message": "The title field must not be more than 1000 characters"
                                                },
                                                {
                                                    "message": "The title field must not be less than 3 characters"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "!! Don`t forget type Bearer before token !!",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}